refactor: Dto 네이밍 수정
